Move components into separate sub-package
